Add: property drawer
