+++ src/LitMotion/Assets/LitMotion.Animation/Editor/LitMotionAnimationComponentDrawer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 44d42c2542ae14559b2ee246182b27ce
